requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
-e git+https://github.com/boulder-python/py-trello.git@b929721397077f31a06603d7bb7843526d2a926b#egg=py_trello
alembic==1.3.0            # via flask-migrate
amqp==2.5.1               # via kombu
billiard==3.6.1.0         # via celery
blinker==1.4
bugsnag[flask]==3.6.0
celery==4.3.0
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
flask-caching==1.7.2
flask-celery-helper==1.1.0
flask-migrate==2.5.2
flask-moment==0.9.0
flask-sendgrid==0.7.1
flask-sqlalchemy==2.4.1
flask-talisman==0.7.0
flask-wtf==0.14.2
flask==1.1.1
gunicorn==19.9.0
idna==2.8                 # via requests
importlib-metadata==0.23  # via kombu
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask
kombu==4.6.5              # via celery
mailchimp3==3.0.9
mako==1.1.0               # via alembic
markdown2==2.3.8
markupsafe==1.1.1         # via jinja2, mako
more-itertools==7.2.0     # via zipp
oauthlib==3.1.0           # via requests-oauthlib
psycopg2-binary==2.8.4
python-dateutil==2.8.1    # via alembic
python-dotenv==0.10.3
python-editor==1.0.4      # via alembic
python-http-client==3.2.1  # via sendgrid
pytz==2019.3              # via celery
requests-oauthlib==1.3.0
requests==2.22.0          # via mailchimp3, requests-oauthlib
sendgrid==5.4.1           # via flask-sendgrid
six==1.13.0               # via bugsnag, flask-talisman, python-dateutil
sqlalchemy==1.3.10        # via alembic, flask-sqlalchemy
urllib3==1.25.6           # via requests
vine==1.3.0               # via amqp, celery
webob==1.8.5              # via bugsnag
werkzeug==0.16.0          # via flask
wtforms==2.2.1            # via flask-wtf
zipp==0.6.0               # via importlib-metadata
